Move calcOnlinePlayers() function to utils; Update method to get playercount in Discord Stats
